requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.2
google-generativeai==0.7.2
python-dotenv==1.0.1
fpdf2==2.7.9
xlsxwriter==3.2.0
rapidfuzz==3.9.3
